séparation de gaze et calibration ( et fin du changement precedent sur le load IHM data
